==> 05-2022 Instacart Basket Analysis/03 Scripts/01. Data preparation - orders and prior_orders.ipynb <==
## Tannis McCartney
## May 26, 2022

### This notebook goes through importing, wrangling, and checking the orders and prior_orders data in preparation for merging with the products dataframe. 

## Contents
### 01 Import libraries
### 02 Import orders data
### 03 Data wrangling orders data
### 04 Consistency checks on orders data
### 05 Change orders data types to reduce memory usage
### 06 Import prior orders data
### 07 Data wrangling prior orders data
### 08 Consistency checks on prior orders data
### 09 Change orders data types to reduce memory usage
### 10 Merge orders and prior orders dataframes
### 11 Export merged dataframe

# 01 Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02 Import orders data

# Turn project folder path into a string
path = r'C:\Users\tmmcc\Google Drive\Data Analytics Bootcamp\4 Python Fundamentals for Data Analysts\05-2022 Instacart Basket Analysis'

# Import orders.csv to df
df_orders = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col=False)
df_orders.head()

# Check shape of df_orders
df_orders.shape

# 03 Wrangling orders data

# Drop the eval_set column from the orders dataframe
df_orders = df_orders.drop(columns = ['eval_set'])

#Rename order_dow column in the orders dataframe
df_orders.rename(columns={'order_dow':'orders_day_of_week'}, inplace=True)
df_orders.head()

# Change data type of orders_id column (in the orders dataframe) to string
df_orders['order_id'] = df_orders['order_id']=df_orders['order_id'].astype('str')
df_orders['order_id'].dtype

# 04 Consistency checks on orders dataframe

# Check for mixed types in the orders dataframe
for col in df_orders.columns.tolist():
    weird = (df_orders[[col]].applymap(type) != df_orders[[col]].iloc[0].apply(type)).any(axis=1)
    if len (df_orders[weird]) > 0:
        print(col)

#### No mixed-type data was found in the orders dataframe.

# Check for missing values
df_orders.isnull().sum()

#### The days_since_prior_order column has 206209 null values. These could be customers who ordered the day the dataset was generated. Others could be customers who have accounts but have not placed any orders yet.

# Create crosstab to compare days_since_prior_order and order_number
crosstab = pd.crosstab(df_orders['days_since_prior_order'], df_orders['order_number'], dropna = False)

# Export crosstab to clipboard for examination in Excel
crosstab.to_clipboard()

#### Where the days_since_prior_order value is 0, the order_number is 1. These are new customers who have only placed one order. A new column to flag these customers as new customers, will be added. A flag of True means days_since_prior_order is NaN because it is a new customer.

# Add new variable to orders dataframe to flag customers who are on their first order.
df_orders_new = df_orders
df_orders_new['new_customer']=np.where(df_orders_new['days_since_prior_order'].isnull(), True, False)
df_orders_new.head()

# Look for full duplicates in the products dataframe
df_dups_ords = df_orders_new[df_orders_new.duplicated()]
df_dups_ords

#### There are no duplicates in df_orders

# Descriptive statistics on df_orders
df_orders.describe()

#### The descriptive statistics seem reasonable.

# Check the shape of orders dataframe after consistency checks
df_orders.shape

# 05 Change orders data types to reduce memory usage

# Check data types and memory usage for df_orders
df_orders.info()

# Change data types for df_orders
df_orders['order_id'] = df_orders['order_id'].astype('int32')
df_orders['user_id'] = df_orders['user_id'].astype('int32')
df_orders['order_number']=df_orders['order_number'].astype('int8')
df_orders['orders_day_of_week']=df_orders['orders_day_of_week'].astype('int8')
df_orders['order_hour_of_day']=df_orders['order_hour_of_day'].astype('int8')
df_orders['days_since_prior_order']=df_orders['days_since_prior_order'].astype('float16')

# Recheck data types and memory usage of df_ordersd
df_orders.info()

#### The memory usage has been reduced from 159.9Mb to 45.7 Mb

# Check the descriptive statistics of df_orders
df_orders.describe()

# 06 Import prior orders

# Import new previous orders data set
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'order_products_prior.csv'), index_col=False)
df_ords_prior.head()

# Check shape of df_ords_prior
df_ords_prior.shape

# 07 Wrangling prior orders dataframe

#### No wrangling is needed for prior orders dataframe

# 08 Consistency checks on prior orders

# Check for mixed types in the prior orders dataframe
for col in df_ords_prior.columns.tolist():
    weird = (df_ords_prior[[col]].applymap(type) != df_ords_prior[[col]].iloc[0].apply(type)).any(axis=1)
    if len (df_ords_prior[weird]) > 0:
        print(col)

#### There are no mixed types in the prior orders dataframe

# Check for missing observations in prior orders dataframe
df_ords_prior.isnull().sum()

#### There is no missing data in the prior orders dataframe

# Look for full duplicates in the prior orders dataframe
df_dups2 = df_ords_prior[df_ords_prior.duplicated()]
df_dups2

#### There are no duplicates in the prior orders dataframe

# Check descriptive statistics on prior orders dataframe
df_ords_prior.describe()

#### The descriptive statistics seem reasonable.

# Check the shape of df_ords_prior
df_ords_prior.shape

# 09 Change orders data types to reduce memory usage

# Check data types and memory usage for df_ords_prior
df_ords_prior.info()

# Change data types for df_ords_prior
df_ords_prior['order_id'] = df_ords_prior['order_id'].astype('int32')
df_ords_prior['product_id'] = df_ords_prior['product_id'].astype('int32')
df_ords_prior['add_to_cart_order'] = df_ords_prior['add_to_cart_order'].astype('int32')
df_ords_prior['reordered'] = df_ords_prior['reordered'].astype('int8')

# Recheck data types and memory usage of df_ords_prior
df_ords_prior.info()

#### The memory usage has been reduced from 989.8Mb to 402.1Mb 

# Check the descriptive stats of df_ords_prior
df_ords_prior.describe()

# 10 Merge orders and prior orders dataframes

# Merge dataframes based on order_id
df_merged_large = df_orders.merge(df_ords_prior, on = 'order_id', indicator=True)
df_merged_large.head()

# Check for a full match
df_merged_large['_merge'].value_counts()

# Remove _merge column 
df_merged_large = df_merged_large.drop(columns=['_merge'])
df_merged_large.head()

# Check shape of df_merged_large
df_merged_large.shape

# 11 Export merged dataframe

# Export data to pkl
df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_combined.pkl'))
