==> 05-2022 Instacart Basket Analysis/03 Scripts/02. Data preparation - products and departments.ipynb <==
## Tannis McCartney
## May 26, 2022

### This notebook goes through importing, wrangling, and checking the products dataframe and departments data dictionary in preparation for merging with the orders dataframe.

## Contents
### 01 Import libraries
### 02 Import products data
### 03 Data wrangling products data
### 04 Consistency checks on products data
### 05 Change products data types to reduce memory usage
### 06 Import departments data
### 07 Data wrangling of departments data
### 08 Convert departments dataframe to data dictionary
### 09 Export products and departments data

# 01 Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02 Import products data

# Turn project folder path into a string
path = r'C:\Users\tmmcc\Google Drive\Data Analytics Bootcamp\4 Python Fundamentals for Data Analysts\05-2022 Instacart Basket Analysis'

# Import products.csv to df_prods
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col=False)
df_prods.head()

# Check shape of df_prods
df_prods.shape

# 03 Wrangling products data

#### No wrangling is needed for the products dataframe

# 04 Consistency checks on products dataframe

# Check for mixed types in the products dataframe
for col in df_prods.columns.tolist():
    weird = (df_prods[[col]].applymap(type) != df_prods[[col]].iloc[0].apply(type)).any(axis=1)
    if len (df_prods[weird]) > 0:
        print(col)

#### The product_name column has mixed-type data. It will be changed to string.

# Change product_name data type to string
df_prods['product_name'] = df_prods['product_name'].astype('string')
df_prods.dtypes

# Find missing observations in products dataframe
df_prods.isnull().sum()

#### There are 16 rows with missing data in the product_name column. The client may be able to provide the product_names for these items.

# Create a subset of the products dataframe to see the rows with missing values
df_nan = df_prods[df_prods['product_name'].isnull()==True]
df_nan

# Copy df_nan to clipboard so client can check product names for these 16 items
df_nan.to_clipboard()

# Look for full duplicates in the products dataframe
df_dups = df_prods[df_prods.duplicated()]
df_dups

# Create a products dataframe that doesn't include the duplicates
df_prods_no_dups = df_prods.drop_duplicates()
df_prods_no_dups.shape

# Describe the products dataframe
df_prods_no_dups.describe()

#### The maximum price is high. It is probably incorrect. This needs to be fixed.

# Show the rows of the dataframe with the high prices
df_prods_no_dups.loc[df_prods_no_dups['prices'] > 100]

# Copy to clipboard
df_prods_no_dups.loc[df_prods_no_dups['prices'] > 100].to_clipboard()

# Turn the incorrect prices into NaNs (NOT best practice in real world)
df_prods_no_dups.loc[df_prods_no_dups['prices'] > 100, 'prices']=np.nan

#### While the two likely incorrect prices have been changed to NaN for now, the correct values should be obtained from the client and the analysis rerun.

# Check results
df_prods_no_dups['prices'].max()

# 05 Change products data types to reduce memory usage

# Check data types and memory usage for df_prods_no_dups
df_prods_no_dups.info()

# Change data types for df_prods_no_dups
df_prods_no_dups['product_id'] = df_prods_no_dups['product_id'].astype('int32')
df_prods_no_dups['aisle_id'] = df_prods_no_dups['aisle_id'].astype('int16')
df_prods_no_dups['department_id'] = df_prods_no_dups['department_id'].astype('int16')
df_prods_no_dups['prices'] = df_prods_no_dups['prices'].astype('float32')

# Recheck data types and memory usage for df_prods_no_dups
df_prods_no_dups.info()

#### The memory usage was reduced from 3.3 Mb to 2.3 Mb

# Check statistics for df_prods_no_dups
df_prods_no_dups.describe()

# 06 Import departments data

# Import departments.csv to df
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col=False)
df_dep.head()

# 07 Data wrangling of departments data

# Transpose departments dataframe
df_dep_t=df_dep.T
df_dep_t

# Add an index to transposed dataframe
df_dep_t.reset_index()

# Create a new header for transposed dataframe
new_header = df_dep_t.iloc[0]
new_header

# Create a new dataframe that starts after the first row of dataframe
df_dep_t_new = df_dep_t[1:]
df_dep_t_new

# Add new header to new dataframe
df_dep_t_new.columns = new_header
df_dep_t_new

# 08 Turn departments dataframe into data dictionary

#Turn transposed departments dataframe int a dictionary
data_dict = df_dep_t_new.to_dict('index')
data_dict

# 09  Export products and departments data

# Export df_prods to pkl
df_prods_no_dups.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'products.pkl'))

# Export the transformed departments dataframe
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
