==> 05-2022 Instacart Basket Analysis/03 Scripts/03. Data preparation - customers.ipynb <==
## Tannis McCartney
## May 26, 2022

### This notebook goes through importing, wrangling, and checking the customers dataframe in preparation for merging with the orders-products dataframe.

## Contents
### 01 Import libraries
### 02 Import customers data
### 03 Data wrangling products data
### 04 Consistency checks on products data
### 05 Change products data types to reduce memory usage
### 06 Export customers data

# 01 Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02 Import customers data

# Turn project folder path into a string
path = r'C:\Users\tmmcc\Google Drive\Data Analytics Bootcamp\4 Python Fundamentals for Data Analysts\05-2022 Instacart Basket Analysis'

# Import customers.csv to df_prods
df_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col=False)
df_cust.head()

# Check shape of df_cust
df_cust.shape

# 03 Data wrangling products data

#### The customer names and genders are not needed for this analysis, and as this information is PII, these columns will be dropped.

df_cust = df_cust.drop(columns = ['First Name', 'Surnam', 'Gender'])
df_cust.head()

# Check the shape of the df_cust dataframe
df_cust.shape

# Check the frequency of the date_joined column - make sure it is more than one value
df_cust['date_joined'].value_counts()

# Rename STATE column
df_cust = df_cust.rename(columns={'STATE' : 'state'})

# Rename Age column
df_cust = df_cust.rename(columns={'Age' : 'age'})

# Rename n_dependants column
df_cust = df_cust.rename(columns={'n_dependants' : 'dependants'})

# Rename fam_status
df_cust = df_cust.rename(columns={'fam_status' : 'household_status'})

# Check column names
df_cust.head()

# 04 Consistency checks on customers dataframe

# Check for mixed types in the customers dataframe
for col in df_cust.columns.tolist():
    weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis=1)
    if len (df_cust[weird]) > 0:
        print(col)

#### The customers dataframe has no mixed-type data

# Check for missing data in the customers dataframe
df_cust.isnull().sum()

#### There is no missing data in the customers dataframe

# Look for full duplicates in the products dataframe
df_dups = df_cust[df_cust.duplicated()]
df_dups

#### There are no duplicates in the customers dataframe

# Describe the customers dataframe
df_cust.describe()

#### The descriptive statistics look reasonable.

# Check the shape of df_cust
df_cust.shape

# 05 Change customer data types to reduce memory usage

# Check data types and memory usage for df_cust
df_cust.info()

# Change data types for df_cust
df_cust['user_id'] = df_cust['user_id'].astype('int32')
df_cust['state'] = df_cust['state'].astype('string')
df_cust['age'] = df_cust['age'].astype('int32')
df_cust['dependants'] = df_cust['dependants'].astype('int32')
df_cust['household_status'] = df_cust['household_status'].astype('string')
df_cust['income'] = df_cust['income'].astype('int32')

# Recheck data types and memory usage for df_cust
df_cust.info()

#### The memory usage has been reduced from 11.0+ Mb to 7.9+ Mb

# Check statistics for df_cust
df_cust.describe()

# Export customer data

# Export df_cust to pkl
df_cust.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'customers.pkl'))
