==> 05-2022 Instacart Basket Analysis/03 Scripts/04 Data preparation - merge orders and products.ipynb <==
## Tannis McCartney
## May 26, 2022

### This notebook merges the orders and products dataframes (that have already been wrangled and checked)

## Contents
### 01 Import libraries
### 02 Import orders and products data
### 03 Merge dataframes
### 04 Export merged dataframe

# 01 Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02 Import data

# Turn project folder path into a string
path = r'C:\Users\tmmcc\Google Drive\Data Analytics Bootcamp\4 Python Fundamentals for Data Analysts\05-2022 Instacart Basket Analysis'

# Import merged orders data created in 4.6 Part 1 and Data Types changed in Part 2
df_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_combined.pkl'))
df_orders.head()

# Check the shape of df_orders
df_orders.shape

# Import products dataset
df_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'products.pkl'))
df_prods.head()

# Check the shape of df_prods
df_prods.shape

# 03 Merge dataframes

# Merge dataframes based on product_id
df_merged = df_orders.merge(df_prods, on ='product_id', indicator = True)
df_merged.head()

# Check for a full match
df_merged['_merge'].value_counts()

# Check the shape of df_merged
df_merged.shape

#### df_merged has 1459 fewer rows than than the orders dataframe. 

# Filter df_orders to show the observations that are omitted
df_test = df_orders[~df_orders['product_id'].isin(df_prods['product_id'])]
df_test.head()

# Check the frequency of the product_id in df_test
df_test['product_id'].value_counts()

# Check for these product IDs in df_prods
# df_prods.loc[df_prods['product_id']==6799]
df_prods.loc[df_prods['product_id']==26519]

#### The missing rows in df_merged are the result of two product_ids missing from df_prods. The client may have this information. For now, nothing can be done.

# Create an object of the missing product ids
# Check the frequency of the product_id in df_test
missing_products = df_test['product_id'].value_counts()

# Copy the missing_products to the clipboard
missing_products.to_clipboard()

# 04 Export merged dataframe

# Remove _merge column
df_merged = df_merged.drop(columns=['_merge'])

# Export df_merged to pickle
df_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
