==> 05-2022 Instacart Basket Analysis/03 Scripts/05. Derviations - orders and products.ipynb <==
## Tannis McCartney
## May 26, 2022

### This notebook derives new columns from the merged orders and products dataframe

## Contents:
### 01 Import Libraries
### 02 Import data
### 03 Derive price_range column
### 04 Derive busiest_day column
### 05 Derive busiest_2_days column
### 06 Derive busiest_period_of_day column
### 07 Export data

# 01 Import libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02 Import data

# Turn project folder path into a string
path = r'C:\Users\tmmcc\Google Drive\Data Analytics Bootcamp\4 Python Fundamentals for Data Analysts\05-2022 Instacart Basket Analysis'

# Import orders_products_merged.pkl
df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
df_ords_prods_merged.head()

# Check the shape of df_ords_prods_merged
df_ords_prods_merged.shape

# 03 Derive price_range column

# Create high-range condition
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range'] = 'High-range product'

# Create mid-range condition
df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range'] = 'Mid-range product'

# Create low-range condition
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range'] = 'Low-range product'

# Check the values in the new column
df_ords_prods_merged['price_range'].value_counts(dropna=False)

# Show the NaN rows
df_nan = df_ords_prods_merged[df_ords_prods_merged['price_range'].isna()]
df_nan.head()

# Check the frequency count of the product_name column
df_nan['product_name'].value_counts()

# Send results to clipboard for inclusion in report
df_nan['product_name'].value_counts().to_clipboard()

#### The NaN price_ranges correspond to the product IDs that were set to NaN in notebook 2 because the prices were incorrect (too high). 

# 04 Derive busiest_day column

# What is the frequency of the orders_day_of_week column?
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna=False)

# Create a for-loop to assign the string value for busiest and slowest days to each row
result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
    if value == 0:
        result.append("Busiest day")
    elif value == 4:
        result.append("Least busy")
    else:
        result.append("Regularly busy")

# Print the result of the for-loop
result

# Add the result to the dataframe as a column
df_ords_prods_merged['busiest_day'] = result

# Check the values in the new column
df_ords_prods_merged['busiest_day'].value_counts(dropna=False)

# 05 Derive busiest_2_days column

# Using frequency count in section 5:
# Create a for-loop to assign the string value for busiest and slowest days to each row
result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
    if (value == 0) or (value ==1):
        result.append("Busiest days")
    elif (value == 4) or (value==3):
        result.append("Slowest days")
    else:
        result.append("Regularly busy")

# Look at the result
result

# Add the result to the dataframe as a column
df_ords_prods_merged['busiest_2_days'] = result

# Check the values in the new column
df_ords_prods_merged['busiest_2_days'].value_counts(dropna=False)

# 06 Derive busiest_period_of_day column

# What is the frequency of the order_hour_of_day column?
df_ords_prods_merged['order_hour_of_day'].value_counts(normalize=True, dropna=False)

#### By setting the normalize argument to true for value_counts(), we can see that 8% of all orders were placed each hour from 10 and 15. In other words, these hours are close to equally busy. 

#### From hours 1-5, less than 1% of orders were placed each hour.

#### For these reasons, I will be grouping the hours for most orders and least orders, instead of just setting one hour for each value.

#### For this column, hours when at least 8% of orders were placed orders will be considered "most orders" and hours when less than 1% of orders were will be considered "least orders." All other hours will be average. 

# Send order_hour_of_day frequency table to clipboard
df_ords_prods_merged['order_hour_of_day'].value_counts(normalize=True, dropna=False).to_clipboard()

# Create a for-loop to assign the string value for most, average, and least orders to each row
result = []

for value in df_ords_prods_merged["order_hour_of_day"]:
    if (value >= 10) and (value <= 15):
        result.append("Most orders")
    elif (value >= 2) and (value <= 6):
        result.append("Least orders")
    else:
        result.append("Average orders")

# Look at the result
result

# Add the result to the dataframe as a column
df_ords_prods_merged['busiest_period_of_day'] = result
df_ords_prods_merged.head()

# Check the values in the new column
df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna=False)

# Check the shape of the dataframe
df_ords_prods_merged.shape

# 07 Export data

# Export dataframe as a pickle file
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_derived.pkl'))
