==> 05-2022 Instacart Basket Analysis/03 Scripts/06. Aggregations - orders and products.ipynb <==
## Tannis McCartney
## May 26, 2022

### This notebook creates flags in the orders_products merged dataset.

## Contents:
### 01 Import Libraries
### 02 Import data
### 03 Create a loyalty flag for existing customers
### 04 Create a spending flag
### 05 Create an order frequency flag
### 06 Export data

# 01 Import libraries

#Import libraries
import pandas as pd
import numpy as np
import os

# 02 Import data

# Turn project folder path into a string
path = r'C:\Users\tmmcc\Google Drive\Data Analytics Bootcamp\4 Python Fundamentals for Data Analysts\05-2022 Instacart Basket Analysis'

# Import orders_products_derived.pkl
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_derived.pkl'))
df_ords_prods_merge.head()

# Check shape of dataframe
df_ords_prods_merge.shape

# 03 Create a loyalty flag for existing customers

# Create a max_order column to determine the number of orders for each customer
df_ords_prods_merge['max_order'] = df_ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Check the results of the above function
df_ords_prods_merge.head(15)

# Set the criteria and execute the code to set the flag for Loyal Customers
df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] > 40, 'loyalty_flag']= 'Loyal Customer'

# Set the criteria and execute the code to set the flag for Regular Customers
df_ords_prods_merge.loc[(df_ords_prods_merge['max_order']  <= 40) & (df_ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular Customer'

# Set the criteria and execute the code to set the flag for New Customers
df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New Customer'

# Print the frequency of new "loyalty_flag" column
df_ords_prods_merge['loyalty_flag'].value_counts(dropna=False)

# 04 Create a spending flag

# Create a mean_prices column to determine the average price across all orders for a customer
df_ords_prods_merge['mean_prices'] = round(df_ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean), 2)

# Check the results of the above function
df_ords_prods_merge.head(15)

# Set the criteria and execute the code to flag Low spenders
df_ords_prods_merge.loc[df_ords_prods_merge['mean_prices'] < 10, 'spending_flag'] = 'Low spender'

# Set the criteria and execute the code to flag High spenders
df_ords_prods_merge.loc[df_ords_prods_merge['mean_prices'] >= 10, 'spending_flag'] = 'High spender'

# Print the frequency of new "spending_flag" column
df_ords_prods_merge['spending_flag'].value_counts(dropna=False)

# 05 Create an order frequency flag

# Create a median_days_since_order column to determine the average price across all orders for a customer
df_ords_prods_merge['median_days_since_order'] = df_ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# Check the results of the above function
df_ords_prods_merge.head(15)

# Set the criteria and execute the code to set the Non-frequent customer flag
df_ords_prods_merge.loc[df_ords_prods_merge['median_days_since_order'] > 20, 'order_frequency'] = 'Non-frequent customer'

# Set the criteria and execute the code to set the Regular customer flag
df_ords_prods_merge.loc[(df_ords_prods_merge['median_days_since_order'] > 10) & (df_ords_prods_merge['median_days_since_order'] <= 20), 'order_frequency'] = 'Regular customer'

# Set the criteria and execute the code to set the Frequent customer flag
df_ords_prods_merge.loc[df_ords_prods_merge['median_days_since_order'] <= 10, 'order_frequency'] = 'Frequent customer'

# Check the results of the above function
df_ords_prods_merge.head(15)

# Print the frequency of new "order_frequency" column
df_ords_prods_merge['order_frequency'].value_counts(dropna=False)

# Check the shape of df_ords_prods_merge
df_ords_prods_merge.shape

# 06 Export data

# Export dataframe as a pickle file
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_derived_aggregated.pkl'))
