==> 05-2022 Instacart Basket Analysis/03 Scripts/08. Data preparation merge orders_products with customers.ipynb <==
## Tannis McCartney
## May 26, 2022

### This notbeok merges the orders_products and customers dataframes

## Contents
### 01 Import Libraries
### 02 Import Orders and Products (with derivations and aggragations) dataset
### 03 Import Customers dataset
### 04 Merge all datasets
### 05 Export data

# 01 Import Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02 Import orders_products_aggragated dataset

# Turn project folder path into a string
path = r'C:\Users\tmmcc\Google Drive\Data Analytics Bootcamp\4 Python Fundamentals for Data Analysts\05-2022 Instacart Basket Analysis'

# Import orders_products_derived_aggragated.pkl
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_derived_aggregated.pkl'))
df_ords_prods.head()

# Check the shape of df_ords_prods
df_ords_prods.shape

# Check the data type of user_id
df_ords_prods['user_id'].dtype

# 03 Import customers dataset

# Import customers.pkl
df_cust = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'customers.pkl'))
df_cust.head()

# Check the shape of df_cust
df_cust.shape

# Check the data type of user_id
df_cust['user_id'].dtype

# 04 Merge datasets

# Merge dataframes based on user_id
df_all = df_ords_prods.merge(df_cust, on='user_id', indicator = True)
df_all.head()

# Check shape of df_all
df_all.shape

# 05 Export data

# Remove _merge column from data
df_all = df_all.drop(columns=['_merge'])

# Export df_all to pkl
df_all.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_all.pkl'))
